:sparkles: feat(comment): comment function all update

+++ src/main/java/com/fis/mylittleboard/domain/comment/service/CommentServiceImpl.java
@@ -5,6 +5,8 @@ import com.fis.mylittleboard.domain.card.repository.card.CardRepository;
 import com.fis.mylittleboard.domain.comment.dto.CommentRequestDto;
 import com.fis.mylittleboard.domain.comment.entity.Comment;
 import com.fis.mylittleboard.domain.comment.repository.CommentRepository;
+import com.fis.mylittleboard.global.jwt.security.UserDetailsImpl;
+import java.util.Objects;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,26 +21,38 @@ public class CommentServiceImpl implements CommentService {
 
   @Override
   @Transactional
-  public void createCommnet(CommentRequestDto requestDto, Long cardId) {
+  public void createComment(
+      CommentRequestDto requestDto, Long cardId, UserDetailsImpl userDetails) {
     Optional<Card> card = cardRepository.findById(cardId);
     if (card.isPresent()) {
-      Comment comment = new Comment(requestDto, card.get().getId());
+      Comment comment = new Comment(requestDto, card.get().getId(), userDetails.getUser().getId());
       commentRepository.save(comment);
     }
   }
 
   @Override
   @Transactional
-  public void updateComment(Long commentId, CommentRequestDto requestDto) {
-    Comment comment = commentRepository.findById(commentId).orElseThrow(() ->
-        new IllegalArgumentException("존재하지 않는 댓글입니다."));
-
+  public void updateComment(
+      Long commentId, CommentRequestDto requestDto, UserDetailsImpl userDetails) {
+    Comment comment = findComment(commentId);
+    if (!Objects.equals(comment.getUserId(), userDetails.getUser().getId())) {
+      throw new IllegalArgumentException("댓글 작성자만 수정할 수 있습니다.");
+    }
     comment.update(requestDto);
   }
 
   @Override
   @Transactional
-  public void deleteComment(Long commentId) {
+  public void deleteComment(Long commentId, UserDetailsImpl userDetails) {
+    Comment comment = findComment(commentId);
+    if (!Objects.equals(comment.getUserId(), userDetails.getUser().getId())) {
+      throw new IllegalArgumentException("작성자만 삭제할 수 있습니다.");
+    }
     commentRepository.deleteById(commentId);
+  }
+
+  private Comment findComment (Long commentId) {
+    return commentRepository.findById(commentId).orElseThrow(() ->
+        new IllegalArgumentException("댓글이 존재하지 않습니다."));
   }
 }
